Compile Fix_1

+++ ghost++/ghost/game_admin.cpp
@@ -45,7 +45,7 @@ using namespace boost :: filesystem;
 //
 
 
-CAdminGame :: CAdminGame( CGHost *nGHost, CMap *nMap, CSaveGame *nSaveGame, uint16_t nHostPort, unsigned char nGameState, string nGameName, string nPassword ) : CBaseGame( nGHost, nMap, nSaveGame, nHostPort, nGameState, nGameName, string( ), string( ), string( ), uint32_t() )
+CAdminGame :: CAdminGame( CGHost *nGHost, CMap *nMap, CSaveGame *nSaveGame, uint16_t nHostPort, unsigned char nGameState, string nGameName, string nPassword ) : CBaseGame( nGHost, nMap, nSaveGame, nHostPort, nGameState, nGameName, string( ), string( ), string( ), uint32_t(), uint32_t() )
 {
         m_VirtualHostName = "|cFFC04040Admin";
         m_MuteLobby = true;
